added new changes

+++ WeatherForecastService/Models/WindWeather.cs
@@ -10,7 +10,7 @@ namespace Weather.Entities.Models
     {
         public string City { get; set; }
         public DateTime Date { get; set; }
-        public int Speed { get; set; }
+        public decimal Speed { get; set; }
         public string Direction { get; set; }
     }
 }
